chore(with-remix-antd): update route convention
