+++ examples/with-remix-antd/app/routes/_auth.register.tsx
