+++ examples/with-remix-antd/app/routes/_index.tsx
@@ -1,0 +1,10 @@
+import { NavigateToResource } from "@refinedev/remix-router";
+
+/**
+ * Since we don't have any routes for the index page, we're redirecting the user to the first resource.
+ *
+ * This can also be done using the `loader` function and `redirect`.
+ */
+export default function Index() {
+    return <NavigateToResource />;
+}
