+++ examples/with-remix-antd/app/routes/_protected.posts.create.tsx
@@ -1,0 +1,49 @@
+import { useForm, useSelect, Create } from "@refinedev/antd";
+import { Form, Select, Input } from "antd";
+
+import { IPost } from "../interfaces";
+
+const PostCreate: React.FC = () => {
+    const { formProps, saveButtonProps } = useForm<IPost>();
+
+    const { selectProps: categorySelectProps } = useSelect<IPost>({
+        resource: "categories",
+    });
+
+    return (
+        <Create saveButtonProps={saveButtonProps}>
+            <Form {...formProps} layout="vertical">
+                <Form.Item label="Title" name="title">
+                    <Input />
+                </Form.Item>
+                <Form.Item label="Status" name="status">
+                    <Select
+                        options={[
+                            {
+                                label: "Published",
+                                value: "published",
+                            },
+                            {
+                                label: "Draft",
+                                value: "draft",
+                            },
+                            {
+                                label: "Rejected",
+                                value: "rejected",
+                            },
+                        ]}
+                    />
+                </Form.Item>
+                <Form.Item label="Category" name={["category", "id"]}>
+                    <Select {...categorySelectProps} />
+                </Form.Item>
+            </Form>
+        </Create>
+    );
+};
+
+export default PostCreate;
+
+export async function loader() {
+    return null;
+}
