+++ examples/with-remix-antd/remix.config.js
@@ -1,6 +1,9 @@
 /** @type {import('@remix-run/dev').AppConfig} */
 module.exports = {
     ignoredRouteFiles: ["**/.*"],
+    future: {
+        v2_routeConvention: true,
+    },
     // appDirectory: "app",
     // assetsBuildDirectory: "public/build",
     // serverBuildPath: "build/index.js",
